Update error status handling to use 32-bit values as per updated RoboClaw documentation

+++ include/roboclaw.h
@@ -100,7 +100,7 @@ class RoboClaw {
                              uint32_t m2_max_distance);
 
   // Get RoboClaw error status bits.
-  uint16_t getErrorStatus();
+  uint32_t getErrorStatus();
 
   // Get RoboClaw error status as a string.
   std::string getErrorString();
@@ -178,7 +178,7 @@ class RoboClaw {
   bool do_low_level_debug_;
 
   // Get RoboClaw error status as a string.
-  std::string getErrorString(uint16_t errorStatus);
+  std::string getErrorString(uint32_t errorStatus);
 
   // Various values are periodically read from the RoboClaw as a group and
   // stored in this structure. Clients wanting these values will get the latest,
@@ -189,7 +189,7 @@ class RoboClaw {
   // method and can be accessed by clients using the g_sensor_value_group_
   // structure.
   typedef struct {
-    uint16_t error_status;
+    uint32_t error_status; // now 32-bit per updated RoboClaw manual
     std::string error_string;
     float logic_battery_level;
     EncodeResult m1_encoder_command_result;
